Lab4/week03-microservices/phase1/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
aiofiles>=23.2.1
httpx>=0.26.0
pydantic>=2.5.3
python-json-logger>=2.0.7
pytest>=7.4.4
pytest-asyncio>=0.23.5
python-dotenv>=1.0.1
